+++ src/context/userStore.ts
@@ -9,8 +9,6 @@ import {
   Wallet,
   WalletParamsWithKind,
 } from '@taquito/taquito'
-import { validateAddress } from '@taquito/utils';
-import { Parser } from '@taquito/michel-codec';
 import { create } from 'zustand'
 import {
   persist,
@@ -19,10 +17,7 @@ import {
 } from 'zustand/middleware'
 import { useLocalSettings } from './localSettingsStore'
 import { NetworkType } from '@airgap/beacon-types'
-import { 
-  getUser,
-  getClaimedDaoTokens
-} from '@data/api'
+import { getUser } from '@data/api'
 import type { RPC_NODES } from './localSettingsStore'
 import {
   BURN_ADDRESS,
@@ -32,12 +27,6 @@ import {
   MARKETPLACE_CONTRACT_TEIA,
   MARKETPLACE_CONTRACT_V1,
   SUBJKT_CONTRACT,
-  DAO_TOKEN_CONTRACT,
-  DAO_GOVERNANCE_CONTRACT,
-  DAO_TOKEN_CLAIM_CONTRACT,
-  DISTRIBUTION_MAPPING_IPFS_PATH,
-  DAO_TOKEN_DECIMALS,
-  MERKLE_DATA_IPFS_PATHS,
   teiaCancelSwapSchema,
 } from '@constants'
 import {
@@ -47,8 +36,6 @@ import {
   createSwapCalls,
   packData,
 } from '@utils/swap'
-import { downloadJsonFileFromIpfs } from '@utils/ipfs'
-import { stringToHex } from '@utils/string'
 import { useModalStore } from './modalStore'
 // import teiaSwapLambda from '@components/collab/lambdas/teiaMarketplaceSwap.json'
 import teiaCancelSwapLambda from '@components/collab/lambdas/teiaMarketplaceCancelSwap.json'
@@ -119,28 +106,6 @@ interface UserState {
   resetProxy: () => void
   /**Transfer tokens */
   transfer: (txs: Tx[]) => OperationReturn
-  /** Votes a DAO proposal */
-  voteProposal: (proposalId: string, vote: string, maxCheckpoints: any) => OperationReturn
-  /** Votes a DAO proposal as representaive */
-  voteProposalAsRepresentative: (proposalId: string, vote: string) => OperationReturn
-  /** Cancels a DAO proposal */
-  cancelProposal: (proposalId: string, returnEscrow: boolean) => OperationReturn
-  /** Evaluates a DAO proposal voting result */
-  evaluateVotingResult: (proposalId: string) => OperationReturn
-  /** Executes a DAO proposal */
-  executeProposal: (proposalId: string) => OperationReturn
-  /** Creates a DAO proposal */
-  createProposal: (title: string, descriptionIpfsPath:string, kind:any) => OperationReturn
-  /** Creates a DAO text proposal */
-  createTextProposal: (title: string, descriptionIpfsPath:string) => OperationReturn
-  /** Creates a DAO transfer mutez proposal */
-  createTransferMutezProposal: (title: string, descriptionIpfsPath:string, transfers: any) => OperationReturn
-  /** Creates a DAO transfer token proposal */
-  createTransferTokenProposal: (title: string, descriptionIpfsPath: string, tokenAddress: string, tokenId: string, transfers: any) => OperationReturn
-  /** Creates a DAO lambda function proposal */
-  createLambdaFunctionProposal: (title: string, descriptionIpfsPath: string, michelineCode: string) => OperationReturn
-  /** Claim DAO tokens */
-  claimTokens: () => OperationReturn
   /** Get the balance for the given address (or current user if not provided) */
   getBalance: (address?: string) => Promise<number>
   /** Mint the token */
@@ -437,298 +402,6 @@ export const useUserStore = create<UserState>()(
             return await handleOp(batch, 'Transfer')
           } catch (e) {
             showError('Transfer', e)
-          }
-        },
-        voteProposal: async (proposalId, vote, maxCheckpoints = null) => {
-          const handleOp = get().handleOp
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          step('Vote DAO proposal', 'Waiting for confirmation', true)
-
-          try {
-            const contract = await Tezos.wallet.at(DAO_GOVERNANCE_CONTRACT)
-
-            const parameters = {
-              proposal_id: proposalId,
-              vote: { [vote]: [['unit']] },
-              max_checkpoints: maxCheckpoints
-            }
-            const batch = contract.methodsObject.token_vote(parameters)
-
-            return await handleOp(batch, 'Vote DAO proposal')
-          } catch (e) {
-            showError('Vote DAO proposal', e)
-          }
-        },
-        voteProposalAsRepresentative: async (proposalId, vote) => {
-          const handleOp = get().handleOp
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          step('Vote DAO proposal as representative', 'Waiting for confirmation', true)
-
-          try {
-            const contract = await Tezos.wallet.at(DAO_GOVERNANCE_CONTRACT);
-
-            const parameters = {
-              proposal_id: proposalId,
-              vote: { [vote]: [['unit']] }
-            }
-            const batch = contract.methodsObject.representatives_vote(parameters)
-
-            return await handleOp(batch, 'Vote DAO proposal as representative')
-          } catch (e) {
-            showError('Vote DAO proposal as representative', e)
-          }
-        },
-        cancelProposal: async (proposalId, returnEscrow = true) => {
-          const handleOp = get().handleOp
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          step('Cancel DAO proposal', 'Waiting for confirmation', true)
-
-          try {
-            const contract = await Tezos.wallet.at(DAO_GOVERNANCE_CONTRACT)
-
-            const batch = contract.methods.cancel_proposal(proposalId, returnEscrow)
-
-            return await handleOp(batch, 'Cancel DAO proposal')
-          } catch (e) {
-            showError('Cancel DAO proposal', e)
-          }
-        },
-        evaluateVotingResult: async (proposalId) => {
-          const handleOp = get().handleOp
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          step('Evaluate voting result', 'Waiting for confirmation', true)
-
-          try {
-            const contract = await Tezos.wallet.at(DAO_GOVERNANCE_CONTRACT)
-
-            const batch = contract.methods.evaluate_voting_result(proposalId)
-
-            return await handleOp(batch, 'Evaluate voting result')
-          } catch (e) {
-            showError('Evaluate voting result', e)
-          }
-        },
-        executeProposal: async (proposalId) => {
-          const handleOp = get().handleOp
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          step('Execute DAO proposal', 'Waiting for confirmation', true)
-
-          try {
-            const contract = await Tezos.wallet.at(DAO_GOVERNANCE_CONTRACT)
-
-            const batch = contract.methods.execute_proposal(proposalId)
-
-            return await handleOp(batch, 'Execute DAO proposal')
-          } catch (e) {
-            showError('Execute DAO proposal', e)
-          }
-        },
-        createProposal: async (title, descriptionIpfsPath, kind) => {
-          const userAddress = get().address
-          const handleOp = get().handleOp
-          const show = useModalStore.getState().show
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          if (!descriptionIpfsPath) {
-            show(
-              'Create DAO proposal',
-              'The proposal description needs to be uploaded first to IPFS'
-            )
-            return
-          }
-
-          step('Create DAO proposal', 'Waiting for confirmation', true)
-
-          try {
-            const contract = await Tezos.wallet.at(DAO_GOVERNANCE_CONTRACT)
-            const tokenContract = await Tezos.wallet.at(DAO_TOKEN_CONTRACT)
-
-            // Initialize the batch operation
-            let batch = Tezos.wallet.batch()
-
-            // Add the token operator
-            const operator = {
-              owner: userAddress,
-              operator: DAO_GOVERNANCE_CONTRACT,
-              token_id: 0
-            }
-            batch = batch.withContractCall(
-              tokenContract.methods.update_operators([{ add_operator: operator }])
-            )
-
-            // Add the create proposal operation
-            const parameters = {
-              title: stringToHex(title),
-              description: stringToHex('ipfs://' + descriptionIpfsPath),
-              kind: kind
-            }
-            batch = batch.withContractCall(
-              contract.methodsObject.create_proposal(parameters)
-            )
-
-            // Remove the token operator
-            batch = batch.withContractCall(
-              tokenContract.methods.update_operators([{ remove_operator: operator }])
-            )
-
-            return await handleOp(batch, 'Create DAO proposal')
-          } catch (e) {
-            showError('Create DAO proposal', e)
-          }
-        },
-        createTextProposal: async (title, descriptionIpfsPath) => {
-          const kind = { text: [['unit']] };
-          return get().createProposal(title, descriptionIpfsPath, kind)
-        },
-        createTransferMutezProposal: async (title, descriptionIpfsPath, transfers) => {
-          const show = useModalStore.getState().show
-
-          for (const transfer of transfers) {
-            const destination = transfer.destination;
-
-            if (!(destination && validateAddress(destination) === 3)) {
-              show(`The provided address is not a valid tezos address: ${destination}`)
-              return
-            }
-
-            const kind = { transfer_mutez: transfers }
-            return get().createProposal(title, descriptionIpfsPath, kind)
-          }
-        },
-        createTransferTokenProposal: async (title, descriptionIpfsPath, tokenAddress, tokenId, transfers) => {
-          const show = useModalStore.getState().show
-
-          if (!(tokenAddress && validateAddress(tokenAddress) === 3)) {
-            show(`The provided token contract address is not a valid tezos address: ${tokenAddress}`)
-            return
-          }
-
-          for (const transfer of transfers) {
-            const destination = transfer.destination
-
-            if (!(destination && validateAddress(destination) === 3)) {
-              show(`The provided address is not a valid tezos address: ${destination}`)
-              return
-            }
-          }
-
-          const kind = {
-            transfer_token: {
-              fa2: tokenAddress,
-              token_id: tokenId,
-              distribution: transfers
-            }
-          }
-          return get().createProposal(title, descriptionIpfsPath, kind)
-        },
-        createLambdaFunctionProposal: async (title, descriptionIpfsPath, michelineCode) => {
-          const show = useModalStore.getState().show
-          let lambdaFunction
-
-          try {
-            const parser = new Parser()
-            lambdaFunction = parser.parseMichelineExpression(michelineCode)
-          } catch (error) {
-            show('The provided lambda function Michelson code is not correct')
-            return
-          }
-
-          const kind = { lambda_function: lambdaFunction }
-          return get().createProposal(title, descriptionIpfsPath, kind)
-        },
-        claimTokens: async () => {
-          const user_address = get().address
-          const handleOp = get().handleOp
-          const show = useModalStore.getState().show
-          const showError = useModalStore.getState().showError
-          const step = useModalStore.getState().step
-
-          step('Claim DAO tokens', 'Claiming Teia DAO tokens', true)
-
-          if (!user_address) {
-            show('Claim DAO tokens', 'You need to sync your wallet first')
-            return
-          }
-
-          // Download the distribution mapping file from IPFS
-          const distributionMapping = await downloadJsonFileFromIpfs(
-            DISTRIBUTION_MAPPING_IPFS_PATH
-          )
-
-          if (!distributionMapping) {
-            show(
-              'Claim DAO tokens',
-              'Could not download the distribution map from IPFS'
-            )
-            return
-          } else if (!(user_address in distributionMapping)) {
-            show(
-              'Claim DAO tokens',
-              'Your wallet is not in the distribution list.\n' +
-                "Sorry, you don't qualify to claim any tokens."
-            )
-            return
-          }
-
-          // Download the file with the user Merkle proofs
-          const fileIndex = distributionMapping[user_address]
-          const merkleData = await downloadJsonFileFromIpfs(
-            MERKLE_DATA_IPFS_PATHS[fileIndex]
-          )
-
-          if (!merkleData) {
-            show(
-              'Claim DAO tokens',
-              'Could not download the user Merkle proofs from IPFS'
-            )
-            return
-          }
-
-          const userMerkleData = merkleData[user_address]
-
-          // Calculate the tokens that the user still can claim
-          const totalTokensToClaim = parseInt(userMerkleData.tokens) / DAO_TOKEN_DECIMALS
-          const alreadyClaimedTokens = (await getClaimedDaoTokens(user_address)) / DAO_TOKEN_DECIMALS
-          const unclaimedTokens = totalTokensToClaim - (alreadyClaimedTokens? alreadyClaimedTokens : 0)
-
-          if (unclaimedTokens === 0) {
-            show(
-              'Claim DAO tokens',
-              'Sorry, but you already claimed all your tokens'
-            )
-            return
-          }
-
-          step(
-            'Claim DAO tokens',
-            'You are allowed to claim ' + unclaimedTokens + ' DAO tokens'
-          )
-
-          // Send the claim operation
-          try {
-            const dropContract = await Tezos.wallet.at(DAO_TOKEN_CLAIM_CONTRACT)
-            const batch = dropContract.methods.claim(
-              userMerkleData.proof,
-              userMerkleData.leafDataPacked
-            )
-
-            return await handleOp(batch, 'Claim DAO tokens', {
-              amount: 0,
-              storageLimit: 400,
-            })
-          } catch (e) {
-            showError('Claim DAO tokens', e)
           }
         },
         collect: async (listing) => {
